requirements.txt
Brotli==1.0.7
dj-database-url==0.5.0
Django==2.1.3
django-cors-headers==2.4.0
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pygments==2.2.0
pytz==2018.7
virtualenv==16.1.0
whitenoise==3.3.1
